Reorganize helpers and model files
Extracted request/response helpers and Route model into their own files
to separate concerns and improve readability of the Otto class. The
Route class was also rewritten to take advantage of logger output and
consistently extend base classes. This improves maintainability and
helps isolate specific responsibilities in the future.
