+++ lib/otto/static.rb
@@ -1,0 +1,36 @@
+
+class Otto
+
+  module Static
+    extend self
+    def server_error
+      [500, {'Content-Type'=>'text/plain'}, ["Server error"]]
+    end
+    def not_found
+      [404, {'Content-Type'=>'text/plain'}, ["Not Found"]]
+    end
+    # Enable string or symbol key access to the nested params hash.
+    def indifferent_params(params)
+      if params.is_a?(Hash)
+        params = indifferent_hash.merge(params)
+        params.each do |key, value|
+          next unless value.is_a?(Hash) || value.is_a?(Array)
+          params[key] = indifferent_params(value)
+        end
+      elsif params.is_a?(Array)
+        params.collect! do |value|
+          if value.is_a?(Hash) || value.is_a?(Array)
+            indifferent_params(value)
+          else
+            value
+          end
+        end
+      end
+    end
+    # Creates a Hash with indifferent access.
+    def indifferent_hash
+      Hash.new {|hash,key| hash[key.to_s] if Symbol === key }
+    end
+  end
+
+end
